funcs/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

joblib==1.2.0
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0